Adjust styles to match desired outcome
I had a rogue center declaration in my sidebar that was messing up the
sizing. Also ended up removing the "end" declaration from the header
links to match as well. Overall I think this is a strange little project
considering that I think I would prefer to use grid for the layout and
then flexbox on the individual sections, but it is good practice
regardless.